Migrate ContextMenu constructor API usages
This CL migrates the context menu instantiations that pass multiple
arguments into using the object form such that we do not have to
maintain the old and the new interface for creating them. I deferred
doing this because I thought there were many, but actually only a few
uses any of the additional options so I thought I might as well upgrade
them all in one go.

Bug: 1242389
Change-Id: Ic39c164a5408c9d98af14aed80c176920a9fe0f1