[e2e] Use host bindings to store UMA events
Instead of firing custom test-only events when a UMA histogram is
recorded, we use the no-op implementation of the corresponding host
binding to instead put the event into a small round-robin buffer. The
e2e tests can then directly look into that buffer and verify that their
expected events are present. This allows us to get rid of the whole
custom event machinery.

R=sigurds@chromium.org

Bug: None
Change-Id: Ief4809bfe0e40c9829909e2bb5d8694784049c3b