Make test error clearer when file is not in ts_library
The default output when Mocha gets an error in its mocharc.js can easily
get lost in all the output - at least, yesterday I was stuck debugging
for ages because I missed the important part of the output in the big
stack trace.

This CL adds more text and a bit of whitespace around the error which
makes it much clearer and easier to spot (see imgur screenshots).

Bug: none
Change-Id: I814936238651e065d1966f31396cf13db1b6bd9b