Improve stepping with source maps
This aligns stepping through source maps with stepping through
Wasm+Dwarf: when stepping from a location, we lookup the corresponding
source location and then make sure we skip all the raw/compiled
locations that correspond to that source location.

To implement the skipping, we reuse the skipList mechanism that is
already used for stepping through Wasm+Dwarf.

Bug: crbug.com/1232347
Change-Id: Ib19695b2078256282189df5b10d9165633849799