Improve Timing reporting for server-delayed requests
This CL stops reporting requests for which DevTools knows
via the RequestWillBeSentExtraInfo event that the request
is not stalled, but for which no response has been received
yet as "Initial connection" instead of "Stalled". This improves
accuracy for web developers, as it indicates the Chrome is
beginning to go to the network. The requestWillBeSentExtraInfo
event doesn't have more information at the moment (in particular,
we don't have timing information for proxy negotiation and/or
DNS resolution, but only that the request job was initiated).

Bug: chromium:1231524
Change-Id: I2e33bc593eed956b00a9fb8a3f507b5a91f2b75c