Migrate ui/legacy/toolbar.css to css.js import
Confusingly, we initially thought this wasn't possible, as the
device mode emulation frame was using this particular CSS file
(transitively). Based on earlier testing, we thought that using
any adopted stylesheet in the device mode emulation frame was
a no-go. That's because Constructed Stylesheets are specced to
disallow moving between associated documents. Since the device
mode emulation frame has a different document than `devtools_app.html`,
it was throwing errors.

However, it now seems to work. That leads me to believe that the
adopted stylesheets are fine, but the actual problem is that the
injectCoreStyles (which we tried to port in https://crrev.com/c/2968288)
won't work and we need an alternative solution for these 4 files.

That said, this should now unblock us to migrate all other CSS
files in ui/legacy, barring the core styles.

R=jacktfranklin@chromium.org
CC=kritisapra@google.com

Bug: 1190991, 1106746
Change-Id: Idcbeb90bff34b82a50749622bfb28859aab72106