Preserve native scrollbar look and feel when using the dark theme on all platforms
DevTools uses the native scrollbar look and feel in light theme, but
custom ones styled via CSS in dark theme. This is because the current
theming architecture styles scrollbars using `::-webkit-scrollbar` CSS
selectors. In contrast to the standard `scrollbar-color` CSS rule, they
have the effect of completely disabling the element's appearance.

It looks like this may have been avoided in the past with the
:not(.platform-mac) selector, however that doesn't seem to be functional
or in use anymore. At the same time, it is questionable if this
complexity is desirable: scrollbars being natively-styled when using the
light theme, custom-styled when using the dark theme only on some
platforms, and natively-styled when using the dark theme on others.

This CL removes this complexity in favor of using the native look and
feel on all platforms, allowing the OS dark/light colors to apply. In
the future, if multiple light or dark themes will become available, this
approach will still work: make sure `color-scheme: dark` is used for all
dark theme selectors, (currently only `.-theme-with-dark-background`) is
used.

Bug: 1244562
Change-Id: I625739ac87a34943842e730f658c465db0b3053b
Signed-off-by: Victor Porof <victorporof@chromium.org>