Display App Id for app manifest
This displays the unique App Id generated from the PWA's manifest in
DevTools.

A short general explainer and a note (in case there is no explicit id
given in the manifest) appear as hover text.

The App Id is only shown if the corresponding chromium flag is turned
on.

Context:
https://groups.google.com/a/chromium.org/g/blink-dev/c/hlxPHfg9GRc

Screenshot: https://i.imgur.com/3cIiXQ4.png

Bug: 1222571
Change-Id: I9305c06a6c49424708668789c3b26234f4a9b638