Introduce AggregationKey abstraction
The IssuesPane currently uses the issue code as key for the map that
holds all active issue views. If we want to support grouping issues
in a more fine-grained way than issue codes, then we need more fine-
grained keys there as well. Ultimately, the decision to aggregate
should be made by the IssueAggregator.

This CL introduces an opaque type AggregationKey that is assumed to
be at least as fine-grained as the issue code and uses that instead
of the plain issue code in IssuesPane.

Bug: chromium:1240622
Change-Id: I4df8c332f40ccec208e90934307f5b50f0a26078