Fix selection color on a11y tree selection
When this file was migrated as part of dark mode a bug was introduced
where we applied the white text color on selection but not the blue
background color.

This CL fixes that and also adds a basic hover state; otherwise the user
gets no feedback when they are hovering over a node.

Fixed: 1247135
Change-Id: I233268e9dc25b8787fff2c65df82420d3369d90a