Removed usages of appendStyle in panels/application
Recently we discovered that some places were using appendStyle directly
to add styles to the page instead of using registerRequiredCSS. A lot
of these calls are in helper functions and so we cannot use the new
registerCSSFiles function as it will not find a document or ShadowRoot
to append styles to. For now we have decided to inline the adoption to
either the shadow root or the document itself.

In another CL I will migrate the remaining usages of appendStyle in
panels/ and rename appenStyle to legacyAppendStyle to prevent it
being used in the future.

Bug: 1106746
Change-Id: I2169726fd028b93085902850434f6627737e75d0