Removed direct usages of appendStyle in panels/application
Recently we discovered that some places were using appendStyle directly
to add styles to the page instead of using registerRequiredCSS. A lot
of these calls are in helper functions and so we cannot use
registerCSSFiles function as getRootNode will not find a document
or ShadowRoot to append styles to. Instead, for now we have decided
to adopt the styles in the closest UI.Widget with a comment explaining
why this decision was made.

In another CL I will migrate the remaining usages of appendStyle in
panels/ and rename appenStyle to legacyAppendStyle to prevent it
being used in the future.

Bug: 1106746
Change-Id: Iefb6e4783b79ed1646a9b1a09723654d915b45a3