Enable efficient rebuilding of Node targets
We had to revert resetting timestamps for all typescript
targets to combat non-determinism in the production build. However,
the Node targets that we have do not suffer from the same fate:
they don't need to integrate with Rollup and don't expose a separate
entrypoint that you need to use types for.

Therefore, opt-in all Node targets to efficient rebuilding. This
should speed up the rebuilding of all test targets under test/e2e.

R=jacktfranklin@chromium.org

Bug: none
Change-Id: If50e7fd25a483825d85c4e235499059f5c1472d9