Fix contrast on delete icon for watch expression
This CL improves the contrast of the delete icon so it can be seen when
you over over the expression. It also tidies up the CSS; we had
min-width being used to add effectively margin to the right hand side so
I've removed that and increased the `right` property.

Fixed: 1248353
Change-Id: I7434641fc577a835bac7ccd0679e2f0d401892de