Specify a single font-family for interaction tests
By default, DevTools specifies multiple font families
and which one will be selected depends on what is available
on the machine. For example, Roboto might be available on
the developer machine but not on the bots. With this change,
we reduce the number of possible font families to one ensuring
that those font families are available on both the bots and
the developer machine.

Bug: none
Change-Id: I18e1ae0924c99ee9bb98c7f258136e80add74273