Use type-safe events for InspectorFrontendHost
This CL adds types for host bindings events that go in the direction of
"Chromium backend -> DevTools frontend". There is already a
similar structure called `EventDescriptors` that fufills a different
purpose. `EventDescriptors` is used to translate the host binding
event parameters, that arrive in an array, to `EventTarget` events,
that are objects with a single "data" property. `EventDescriptors`
doesn't contain any type information for the event parameters and
it is used at runtime. The new `EventTypes` type on the other hand
only contains type information and is not used at runtime.

Please note that `EventTypes` can't be used to check the "dispatch"
side, as the event translation + dispatch happens in a dynamic
fashion.

R=pfaffe@chromium.org

Bug: 1228674
Change-Id: I1dbe6ba6b0a494963bf443f25dedc860fa3e28e1