Add column number to linkified sources
This changes updates the Linkifier to include the column number in the
link text if available. Also taking the opportunity to remove some
run-time type checks that are no longer needed thanks to the TypeScript
migration.

Including the column number in source links helps developers who are
looking at a performance profile "offline" find the function that was
called within the source file. When only a line number is given, this
is more difficult, especially for minified JavaScript files where there
may be many function definitions on the same line.

Bug: 1252049
Change-Id: I1e33583010c8f5d62789e4341ca5b4f6b2c9cd58