Be consistent when displaying async stack traces.
In case of async functions, V8 captures an additional frame from the
resumption after an `await` site, which was explicitly removed from the
display in the Call Stack, but this cleanup wasn't applied to async
stack traces shown in the Console.

That lead to the inconsistency observed in crbug.com/1229962, where it
looked like it was pointing to the wrong `await`, when in reality it
was pointing to where the async function resumed last (an implementation
detail of async functions if you will).

With this change we now consistently clean redundant frames from async
stack traces also for Console messages, so that we display the same
frames in Call Stack and Console.

Drive-by-fix: Also use the same divider text for async functions - the
'await in $fun (async)' text - in both Call Stack and Console.

Before: https://imgur.com/3y2KLED.png
After: https://imgur.com/i9tb8fb.png
Bug: chromium:1229962
Change-Id: Ia206c83d7d5534e21c440d0857fc64d72a1331f4