Add messages for updated cookie size checks
We recently made some changes to the way that Chrome handles cookies
deemed to be too large. Specifically, we modified Chrome's existing
cookie size check to more closely align with what other browsers do
(see [1] for more details), and then applied this check more
consistently across other cookie interfaces (for example, the web
request extensions API) behind an enabled-by-default Finch flag.
For reference, [2], [3], [4], [5], [6], and [7] are the CLs that
implement these changes in Chromium and in the corresponding WPTs.

[1] https://groups.google.com/a/chromium.org/g/blink-dev/c/0N5BePVCPVo/m/nHAhSXf3AwAJ
[2] https://chromium-review.googlesource.com/c/chromium/src/+/3104225
[3] https://chromium-review.googlesource.com/c/chromium/src/+/3106829
[4] https://chromium-review.googlesource.com/c/chromium/src/+/3111249
[5] https://chromium-review.googlesource.com/c/chromium/src/+/3145752
[6] https://chromium-review.googlesource.com/c/chromium/src/+/3097489
[7] https://chromium-review.googlesource.com/c/chromium/src/+/3155503

Also, we sent the following document to DevRel for their thoughts:

https://docs.google.com/document/d/1p4hJIcmwW7MiGU9eAwYWSO2iovWlIvVmWwJvnNG4OKw/edit?usp=sharing&resourcekey=0-Og1pEfRn9TzLg_KK3etOnw

Bug: 1225342
Change-Id: Ie5e9ab39927027de3b609973df2d3f23ab66dfb6