[ux] Render async stack traces consistently across DevTools front-end.
This is a small follow up change to https://crrev.com/c/3193535, which
adjusts the V8 Inspector to report async stack traces for async
functions (`await`s actually) to be consistent with other async tasks,
meaning that the async part contains only frames leading up to the task,
but not the task itself. With this change we also use the 'await in foo'
string consistently across DevTools front-end, in particular the async
stack trace in the Call Stack and the Console now looks consistent.

Screenshot: https://imgur.com/9ldRoLo.png
Fixed: chromium:1254259
Bug: chromium:1229962
Change-Id: Icdc7343e7df49afd05854c4324e359f5b68bb8c6