[cleanup] Remove the 'Empty sourcemap auto-stepping' experiment.
With https://crrev.com/c/3172834 we previously aligned the stepping
behavior of sourcemaps (both JavaScript and WebAssembly) with that
of DWARF (WebAssembly only), but we still gated that adjustment
behind the 'Empty sourcemap auto-stepping' experiment.

This enables the aligned behavior by default and removes the (confusing)
and according to usage metrics almost completely unused experiment.

Bug: chromium:1232347
Change-Id: Idfa62a2cfa337d4ced3486baa3249cb95a8d8c16
Fixed: chromium:1018234