[Accessibility]: Set title for grid/flex overlay color picker
Previously the input for activating the color picker for grid/flex
overlays in the Layout pane of the Elements panel simply read "text"
when focused while using a screen reader.

This change adds an explicit title similar to other nearby inputs which
reads "Choose the overlay color for this element".

Bug: 1255963
Change-Id: I354d67d049992713733d6fef487ca1c4788346a7