[sync] Add syncable storage to Settings instance
This CL adds a "syncStorage" that mirrors "globalStorage" and
"localStorage". The corresponding "StorageType" enum variant will
be introduced in a follow-up CL. As the syncalbe storage is very
closely related to the global storage, the same backing store is
used for both.

This Cl also adds a rudimentary unit test for the Settings class.

R=dsv@chromium.org

Bug: 1245541
Change-Id: Ibd9d2e5c63b6bea0f909d901c8506b666ef2d673