Check UILocation when adding/removing breakpoint decorators
This change updates the DebuggerPlugin class's decorationByBreakpoint
map to key on both Breakpoint and UILocation when adding and removing
breakpoing decorators in the UI. Before this change, if a Breakpoint was
associated with multiple UILocations, then attempting to remove that
breakpoint would not update all of its decorators correctly. This is
because the first breakpointRemoved event that was received would
remove the Breakpoint entry from decorationByBreakpoint, and then any
subsequent breakpointRemoved events for the same Breakpoint with a
different UILocation would have no effect.

After the change, breakpoints with multiple UILocations and decorators
are updated correctly. An e2e test has been added to verify this.

Bug: 1251675
Change-Id: Ia33593af5a3efcb08cf219d2e21932406760f3fd