user agent client hint form focus order bug fix
The user agent client hint (UACH) form is used in network condition
panel and settings/device page. When user deletes a brand row in the
UACH form, if there is no existing brand then keyboard focus should
land on the "Add Brand" button or else it should land on the next brand.

Bug: 1256936
Change-Id: I6b4111665d0771be51d648ba625c74abe0cec626