Fix MacOS Hover Text display for infobar 'Learn more' link
The 'Learn more' text was styled as a link, but didn't have a link role
which is why the Hover Text feature displayed it as text.

Fixed by adding role='link' to align the role communicated to a11y
tools with the element's visual representation.

Bug: 1257341
Change-Id: I30c6229fcdb228934840168342417c2b88c84051
Before: https://imgur.com/a/djbWgpZ
After: https://imgur.com/a/c9y3pzn