Fix <length> component editing issues
This CL fixes two issues:
1. When the mouse clicks on a <length> value and releases without
moving, the "mousemove" event listener bound on the document won't be
removed, causing other "mousemove" events to be blocked.
2. When the mouse clicks on a <length> value, starts moving, and stops,
it won't reset the Styles pane's editing state, which in occasions
would cause the Styles pane to stay in this "editing" state and won't
accept other modifications.

Bug: chromium:1249902, chromium:1245431
Change-Id: Ib651da50514e64fe616d5b330fb3980f1e2a0d1d