Fix broken paths in `NativeFunctions.js` generator
`scripts/javascript_natives/index.js` expects `NativeFunctions.js` to
reside at a specific path, and broke when a CL [1] moved it to another
location, and broke (in another way) when the DevTools front-end
codebase moved to a separate repository. Similarly, the script expects
a Chromium checkout at a specific location relative to itself, which
hasn’t been correct since the DevTools front-end codebase moved to a
separate repository.

This patch updates the expected locations, but does not re-generate
the output since the script requires additional follow-up fixes.

[1]: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2814662

Bug: chromium:1255619
Change-Id: I066fb4349f045df5a12edd1378d7acdaa733a61b