Add shortcut to search panel for expanding/collapsing all files
See the feature request in crbug/1255073.
Originally, the feature request is only for
expanding all the files.
Since the implementation is trivial and the
shortcut is already ready for collapsing
functionality I decided to add it as well

Though if you think it is unnecessary, I can remove it @bmeurer :)

I have selected "Ctrl + Shift + }" for expanding
and "Ctrl + Shift + {" for collapsing.
The reason is, there isn't a common shortcut for
expanding or collapsing all blocks/search items in
IDEs (I checked Webstorm and VSCode). However,
VSCode uses "Ctrl + Shift + }" for expanding
specific code blocks. Since we handle expanding
tree items only with right arrow, this seemed like
a good candidate for expanding all the items. WDYT?

One other decision was to add the shortcut to the
whole search panel (including the text box)
thinking that it might be useful for users even
when they are focused on the text box:
* Say, they have written "min"
* They expanded all the files without leaving focus
* Then wanted to search for another term and do the same.
So, this way, they can do it while focused on the text box.

So, overall, I have these questions:
* WDYT about the chosen shortcuts?
* WDYT about adding the shortcut to the whole search panel?
* Is there a way that we can make this shortcut
known to the users? (Maybe there is a shortcuts
document somewhere that we can add?)
* WDYT about the request of expanding all the
occurrances as well? Should we also do that with these shortcuts?

Thank you! :)

Bug:1255073
Change-Id: I4e8ff3cf82b532346acb1ac6c247f0f361477287