Do not act on forwarded resume action
Previously we would forward F8 events on the target page
to the SourcesPanel. This would trigger two separate resume actions,
one in the inspector overlay, and one in DevTools front-end. The
forwarding is still required, as we are tracking the user metrics.

This adds a check to the SourcesPanel to only act on the resume,
if it's not coming from the forwarding, and if the overlay is enabled.

Fixed: chromium:937130
Change-Id: I17bebd95b53d9c8e802a5b2a0320e7b4225f144d