Fix usage of private field in TracingModel
The indirection of `Sorter` was causing issues, where the cast to `any`
was masking over the fact that the `Sortable` interface was intended to
be referencing `NamedObject` instead. Since TypeScript allows us to
specify a generic bounded type, we can move the static method to
`NamedObject` instead and cleanup the redundant interfaces and classes.

R=alexrudenko@chromium.org

Bug: 1172300
Fixed: 1082064
Change-Id: I1c6b6d22edc8247e222ae6ecdef568d229724f8d