[settings] Refactor settings rendering code
This CL refactors how the sections of the "General Settings" tab are
built. The existing code does 2 passes:
  1) It creates a container for each setting category
  2) It iterates all settings appending them the the correct container
     depending on the settings' category

The new code builds each section in one step by filtering all
settings that will end up in that section.

This refactoring is necessary to allow an easier transition to
web components. Instead of incrementally adding settings to a
component causing many re-renders, we can initialize and render
a future section component in a single step.

Drive-by: Rename `addSettingUI` to `addExtensionsSection` given
what the code actually does.

R=alexrudenko@chromium.org

Bug: 1245541
Change-Id: I93e2b68dbb945a3ede5a8ded9da5ad88fd434fb5