Conditionally unset tree focusable in Events
When no event handlers are present for a given element, the focusable
tree outline would appear to be a line in tab-order. This change
corrects that issue by unsetting the tree's focusable state whenever
the tree has no elements (specifically when the "No event listeners"
message is visible).

After: https://imgur.com/a/r0TDY4n

Bug: 1171703
Change-Id: Iae79239c4c95ea5561d51515fe09062242e03835