Fix unbalanced quotes from being sent to the inspector
This CL fixes the problem where if we enter a quote in a CSS property
value (and this quote is unclosed at the moment), the next CSS
declaration would be removed after the edit. The chain of events are:
1. an unclosed quote is entered in the CSS property value;
2. DevTools supports live editing, sending this unclosed quote (hence
   erroneous propety value) to the backend to update the style text;
3. since the style text is updated per declaration block instead of per
   single declaration, this unclosed quote will cause the parser to
   "hoist" the next CSS declaration to be part of the modified (wrong)
   declaration;
4. the result will be sent back immediately to DevTools frontend, and
   when the remote (backend) state is updated in the frontend, the
   "hoisted" next-line declaration will be shown in the first line.

This fix balances the unclosed quote in the property value, effectively
stopping the parser from treating the next line as part of the property
value.

Fixed: 792396
Change-Id: I7da6f6459414fc27dc8f47060b0474d66b20d10d