Use case sensitive look up in build scripts to copy file contents.
These scripts check if a file exists and whether its content is
different from a source file before copying it. However the file look up
done with fs.existsSync() could be case-insensitive depending on the
underlying file system.

This CL implements the a case-sensitive file look up to compare against
the exact path name of the file being copied.

Bug: none
Change-Id: I0dc175c2292b3caea541a585562e8e25db637340