Add loadLegacyModule on Runtime for browser tests
Browser tests use their separate test framework in Tests.js. This
specific test requires the full panel to be loaded, which isn't
necessarily true when the JavaScript function is executed.

To combat flakiness, the previous version of this test was explicitly
loading the elements panel using the legacy `loadModulePromise` method.
Instead, we should directly import the module. However, since `Tests.js`
runs in a different scope, we can't use relative imports there and we
have to let the runtime take care of that.

R=caseq@chromium.org

Bug: 1266640
Change-Id: If3448984c1f903d2503e8127f52f1385d95959ca