[core] Remove debug-only setTimeout().
When using --custom-devtools-frontend (or other means that trigger
'debugFrontend' parameter to be set to true), dispatching of browser
messages is delayed via an additional `setTimeout`. This however causes
trouble for example with context menu selection events, where it races
with the removal of the menu item event listeners.

Fixed: chromium:1266826
Change-Id: Idcb0ba5b19854cefcf5a3f8ca2bb2a617ec047e9