Create generic PanelFeedback component
This CL takes the "Preview feature" that is already used in the CSS
Overview and Recorder and makes it generic, as we'll want to use this
component whenever we are launching a new panel.

It takes one property, which is the URL the inline "Feedback" link
should link to, and takes a slotted element for the "Quick start" link.

I do plan to add some tests to this component before landing, but asking
for review now to get opinions on the API design before diving in.

I also have not included the small "Feedback" button that sits in the
lower right corner - I think that should be its own component and I will
do that in a follow up CL also.

Bug: 1268802
Change-Id: I6c179663d8d10f6671c48d27d5a25da6ffdd9897