Enable Testrunner to trim multiple spaces down to a single one
After adding icons which function as links to console messages in [1],
some tests were failing [2] because the icon components contain multiple
text nodes of a single space character each. These spaces were being
rendered in the test output. This CL adds options to reduce multiple
consecutive spaces down to a single one for test output. The tests
are being turned on again in [3]

[1] https://crrev.com/c/3269837
[2] e.g. https://ci.chromium.org/ui/p/chromium/builders/ci/
  Linux%20Tests%20(dbg)(1)/100288/test-results
[3] https://crrev.com/c/3275975

Bug: 1260776
Change-Id: Id352e137fa2b8d153b3c8031f1c44d8aedb46c8d