Move legacy file loading to legacy_test_runner
DevTools startup has slowed down a bit over the last year. The main
suspect that we have is the amount of files that are loaded at startup.
We have historically made significant improvements by removing file
loading out of the critical path, for example by removing the `module.json`
files from various modules. These have all attributed to improvements.

For our shell, we were still loading these legacy files. However, now
that we are fully type-checked with TypeScript, we have for the most part
removed our reliance on globals. There are only a handful of places where
we still rely on globals, most notably `MainImpl` for instantiation of
singletons and a single declaration in `ui/legacy/Panel.ts` which sets
values on `UI.panels`. Other than that, there don't appear to be anymore
usages of globals.

Therefore, we can move these legacy files out of our main shell and into
the legacy_test_runner. Doing so, our production build will now include
fewer files and hopefully also boot faster, while we still load the
legacy files for our layout tests (which still rely on globals).

It isn't clear what the performance impact will be, but cleaning this up
is good and we are relatively confident this will alleviate at least
some of the performance problems that have been reported.

R=yangguo@chromium.org

Bug: none
Change-Id: Ie34fbf6b82f3d1ba11bed623a9cc3d2a03c4c0c7