[cleanup] Remove obsolete use of `Runtime.CallFrame.url`.
In the `CallStackSidebarPane`, we have use `Runtime.CallFrame.url` as a
fallback in case the `DebuggerModel` for a given async stack trace
cannot be determined. But this is actually dead code, since we will stop
iterating the async stack in case the `DebuggerModel` cannot be found;
it was just that the code didn't express this clearly.

This CL restructures the code to express this clearly and removes the
dead code which was using `Runtime.CallFrame.url`.

Bug: chromium:1270248
Change-Id: Ia51a7517134dc8eba6a0441d9b690a884ddef8aa