[linkifier] Always favor script IDs over source URLs.
Previously `Linkifier.maybeLinkifyScriptLocation` had a weird fallback
where even when a script ID is presented, it'd go hunting through the
list of known scripts by source URL if the script ID didn't resolve to
an existing script. This however is wrong (an example of this generally
can be found in crbug/1270248), since URLs don't uniquely identify
scripts and we might send the developer to the completely wrong script
(that by chance has the same URL) instead of indicating that we don't
know anything about the script yet/anymore.

This CL by itself doesn't fix crbug/1270248 yet, since the rest of the
linkification logic will still fall back to using source URL, but at
least we no longer hook up to arbitrary wrong scripts.

Bug: chromium:1270248
Change-Id: I61dd2d9346bf25ae72e0400706324e924f1c12ef